+++ test/tests/04 - database_collection/03 - _finalize.php
@@ -1,17 +1,17 @@
 <?php
 
 /**
- * Testing _database_collection_finalize().
+ * Testing _finalize().
  * php version 8.0.0
  *
  * @author Bourg, Sean <sean.bourg@gmail.com>
  */
 
- use function pirogue\_database_collection_finalize;
+ use function pirogue\database_collection\_finalize;
 
 require_once(implode(DIRECTORY_SEPARATOR, [_PIROGUE_TESTING_PATH, 'include', 'pirogue', 'database_collection.php']));
 
-pirogue_test_execute('_database_collection_finalize()', function () {
-    _database_collection_finalize();
+pirogue_test_execute('_finalize()', function () {
+    _finalize();
     return empty($GLOBALS['._pirogue.database_collection.connections']) ? '' : 'registered connections exist.';
 });
